Add Command class
The Command class primarily serves as a wrapper of the raw received JSON
data. It also supports equality checks, and contains class 'constants'
used to check for the type of command received.

Also enable future of join_game() such that it is actually called when
a JOIN command is received.