Add JSON object value getter and memory checks
This getter acts as a wrapper around the json-c function, executing a
check on the result and raising an error if the function is
unsuccessful.

Also add memory allocation checking for hash function, and initialise
pointers to NULL. I feel this increases robustness of the code without
having non-specified initial values with undefined behaviour.