Add JSON to card method, more card tests
Tests now test card -> JSON -> card returns a card with the same values.
Also improved the manner in which tests are conducted, and added an okay
coverage of all valid card values. Would be good in the future to add
more invalid values which should break (this is a good thing). Maybe
return a NULL from card_new() which the user has to check for? This
would allow a test to be written which would not exit the program via a
call to err().

Decided to rename card_to_JSON() to JSON_from_card(), and similarly with
its counterpart card_from_JSON(). I believe this naming makes it more
apparent the function returns a newly allocated JSON object or card
structure rather than relying on pre-allocated structs.