Fix multiple command sends creating invalid JSON
If the last client needed to start a match were to join, it receives a
JOIN and START command in the same packet. However, as the packets are
simply JSON objects as strings, parsing errors were raised as there
was no delimiter between the two JSON objects.

To fix this, a command queue was added to the client structure which is
able to be populated with various commands using enqueue_command(). The
queued commands are finally sent to the client by popping commands off
the queue and adding them to a JSON array which is packed into an object
and sent to the client through send_commands_to_clients().

There may be some memory issues with this approach, as memory allocated
to the JSON objects representing commands is never explicitly freed.
However, the hope is that a freeing call on the parent JSON object will
also free memory associated with its children. This may not be the case.

Alongside this, the python client has also been updated to handle a list
of commands being received, rather than the one command object as was
happening before.

The billionaire_join() function now no longer generates an id, the user
must explicitly provide an id to this function. This way feels better.