Add FINISH command, thrown when not enough players
When the number of players drops below the player limit, the server will
now issue a FINISH command to each client. The clients must obey this
command and cease all playing immediately.

As a consequence, send_commands_to_clients() now checks if the command
queue for a client is empty before it attempts to construct a JSON
packet.