Use correct deallocation for json_object*
Using free() was a misnomer. I have also had to reduce the scope on a
couple of variables in billionaire-server.c. I should really run
cppcheck over the entire source directory to check it all adds up.

I have also removed calls to free() on the string representation of JSON
commands, as I saw in check_card that this lead to an incredible amount
of memory being freed, which I believe suggests that I am implicitly
calling free twice on the same pointer.