Add memory testing program mem_test
This program can be run through your favourite memory checking routine
to check for any leaks.

This program should feature a large coverage of the many different ways
memory can be implicitly and explicitly allocated, and of course the
correct ways in which they should be deallocated to prevent memory
leaks.