Move bot logic into base class BaseBillionaireBot
This base class is the class each actual bot must inherit from. It deals
with handling server connection logic, as well as responses to server
commands that are compliant with expected behaviour. Under normal
circumstances, it should not be necessary for an end user to have to
make changes to this base abstract class. All user-defined functionality
can be accomplished by creating a new derived class.

BaseBillionaireBot uses a new meta class BotMeta to decorate the methods
of its derived classes to enforce the correct server compliance. For
example, currently the issue_command() method is wrapped to ensure that
commands sent from it will be wrapped in a CommandList if not already.