Truncate overflowing data into command string
Previously, was simply printing the data which lead to overflows if one
client's command was smaller than another's. Now, snprintf is used on
the data to copy only the valid part of the command.

Notice that this does not mean the entirety of the data is read. This
method assumes that received data is less than 8kB in size.