Add Cards container object
This Cards object needs to have wide-reaching and diverse functionality
in order for it to be used by the end-user to devise competitive
trading algorithms. It must support basic querying of the number of
cards, and the type and associated quantity. It must also support a
straightforward way of extracting multiple cards of the same type which
is essential to forming ASKs.

As a consequence, the Card object is now fully hashable, and supports
conversion between a Card ID, which is the least amount of information
that is needed to know everything about a card. For example, if you
know a card is a BILLIONAIRE, you know it doesn't have a commodity type.
Likewise, if you know a card is DIAMONDS, you know it is a commodity.

There will need to be a test suite written for both card classes to
maintain operability into the future.