Implement Cards into bot base, better defaults
The Cards container object is now used for the self.hand field, instead
of the simplistic array of Card objects that it was using before.

Better defaults have also been added so that the program doesn't crash
when accessing things for the first time, i.e. when issue_command() is
called during Future setup.