Add simple ASK algorithm
This particular algorithm asks for the maximum number of cards it has in
the hand. Just a proof-of-concept to demonstrate some Cards methods.

There is a slight issue in that the bot sends an ASK command with no
cards in it before it begins going through the cards in its hand. I feel
like it is an issue with the asyncio.Queue used with commands, and how
the issue_command() is called once before everything begins, but I
could not tell you as the queue seems to be empty when the bot JOINs
and STARTs the game. Something to keep in mind, I guess.