Add support for command parsing and processing
Includes the start of the actual game logic, as contained in the server.

In the future, should be able to make an array-extraction method just
to parse command lists sent by clients. Issue before was needing to
deallocate the parsed JSON object without also deallocating the array
object that was extracted from this. This can be done with a simple
_get() on the resulting array object, and a _put() on the full parsed
object. This should increment and decrement the reference values such
that only the parsed object is freed, which means the user does not have
to worry about deallocating it later on when they no longer have a
reference to it!