Add start of better asyncio handling for commands
In the future, would like to form a routine that can dynamically react
to server messages, and make judgements on what actions to take itself.
It needs to form a part of a generic framework that is primarily easy
for an end user to use and create their own Billionaire tactics without
having to know anything about the intermediate protocols, nuances, etc.