Add card_location structure and associated tests
This structure will eventually replace the previously-used card** and
accompanying size_t* arrays.

The structure supports initialisation by an arbitrary list of cards, as
well as incremental addition of cards. The structure dynamically
increases the size of the array used to store the cards, whose size can
be fixed at the current number of cards with a call to
fix_card_location_size().

Eventually, this structure will need to support constant-time deletion
of cards, as well as popping of cards. These will be needed in the
future to change player's hands stored in the game_state structure used
to check the validity of a given ASK by a player.