Reimplement card_location to use card counts
This implementation removes the need to store card structures entirely.
Now have a reduced memory footprint, and very efficient storage and
retrieval of an arbitrary amount of cards from any location.

A card_location only counts how many of each type of card it has. Hence,
cards can be added and removed by specifying a card_id and an amount. A
card_id is a unique enumeration defining a card, and is either the type
of commodity a card is, or the type of special card it is.

Adding and removing cards is constant-time, as each card_id corresponds
to an entry in an array, which can be increased or decreased at will.

In the future, will need to implement integration with errno.h to allow
proper error handling in cases where too many cards are told to be
removed from a card_location.