Replace card functionality with card_location
This is a large overhaul of the code to remove the use of card structs.
They were made redundant with the reformulation of the card_location
struct to instead count the instances of a particular card, rather than
keep references to each card sequentially.

A new struct was required which would preserve the order of cards
inserted into it. This particular property is paramount to the shuffling
of a group of cards and subsequent dealing to other players. The struct
is the card_array, and features in card_location.h as essentially a
helper struct which should only be used to shuffle a deck of cards
before they are dealt (see game_struct implementation for instance).

Have fully integrated the card_location changes into the server. The
Python client is yet to be rewritten to use these changes, and as such
should not be used in its current state.

Tests have also been written for the extensions to the card_location
struct, including the addition of the card_array struct. This commit is
yet to feature a test for the shuffling routine, so be aware of this
going forwards.