Fix pointer warning in JSON_ARRAY_FOREACH
This warning was a result of me not knowing how to properly initialise
a for loop with multiple variables.

Would like a way of limiting the scope of the variables used in the loop
to just the loop scope itself, but then braces are non-intuitive when
the macro is actually used. Something to think about.