Add parsing routine for inbound command lists
All received command lists are now parsed with a single routine which is
leak-proof. An example of such a use has also been added to mem_test.

Placeholder NEW_OFFER parsing has been added to showcase how to pick
apart the command structure. This will be changed in a future commit to
implement the correct behaviour.