Split card_array struct into separate header
The program should have each struct object contained in its own header
file. In the case of card_array, its methods are used to generate
a card_array from a card_location, and then shuffle the card_array to
generate an array of card_location structs.

Although it is primarily a helper class, it should be in its own files
to maintain ease of extensibility into the future.