Add handling for EOFFEROVER, clean cancel handling
Errors resulting from the call to cancel_offer() are now handled in the
same manner as all other errors. Previously, the check was made on the
return value of cancel_offer(), now the value of cmd_errno is checked
explicitly. This will hopefully prevent any issues in the future when
the behaviour of cancel_offer() may be changed so that a return value of
NULL is valid.