Implement three missing commands
The three commands implemented were SUCCESSFUL_TRADE, CANCELLED_OFFER
and BOOK_EVENT. No testing has been done on these, so expect some bugs.

Also removed the magic number for the maximum number of participants in
a book event. Of course, it's obvious that there can only be a maximum
of two participants in such an event, as a trade is only ever between
two players. However, it is much better to be explicit with the coding,
and also keep the constant defined in just one place.