Add rudimentary trade/cancellation support
Note that in the case of a trade successfully being made, the trade is
not yet completed in the current code. This is because the other client
(the one whose trade in the book was successfully matched against the
incoming trade) is only accessible by ID. This does not return a client
structure which is needed to send commands to. Hence, there needs to be
some sort of hashing implemented for clients so that given a client ID,
a client structure can be found and used to send material to.

In the worst case, a simple O(n) check can be done on the client queue
to find the matching client, but I am more interested in the elegant and
quick solutions that involve a lot more effort but work much more
smoothly.