Add client.o to build targets
Forgot to add this to 595b1e2 all those commits ago. The reason why this
took so long to add is that when working on my local machine, I use a
different set of optimisation flags to build the server. The opt flags
given in the repo are intended to work on Raspberry Pis, so it is
perfectly normal for me to commit changes when my Makefile has been
modified. In this case, I had forgotten that I had modified the Makefile
to also build the client.o object file.

But now it's rectified and there's nothing you can do that will make me
look bad. Except point out how I clearly don't know how to use any CI
suites and if I did this would not have been a problem.