Change del_client to use client struct over ID
This allows for a stronger check that the client found indeed matches
the client requested for deletion, on the off-chance that the client IDs
collide. It also makes more sense, as the only use case for this
function is when a client disconnects, and we always have access to the
disconnecting client object.