Add internal client hand state
This state will be updated every time a command is sent to the server
that modifies the client's hand. This will be checked against for new
offers to ensure that the offers put in the book did come from the
client's hand.

The hand is not allocated during client initialisation, as it will be a
pointer to the delt hand sent to the client at the beginning of the
game.