Add offer subset checking method
This method will be called as part of a checking suite to ensure the
offer given to the server by the client is a valid one. This check
ensures that the cards given in the offer came from the client's hand,
or equivalently that the offer is a subset of the hand.