Move all offer validation into one method
This method replaces the subset method defined previously, and makes a
variety of further checks on the offer:
 - Contains non-zero amount of cards
 - Contains a tradable amount of cards
 - Is a subset of the client's hand
 - Contains <= OFFER_MAX_UNIQ_COMMS unique commodity types
 - Contains <= OFFER_MAX_UNIQ_WILDS unique wildcard types

The method does not check for INVALID card types, because the
card_location struct does not store INVALID card types. This means any
INVALID cards given to the server are completely ignored. This is "good
enough" behaviour for the server, as it seems unnecessary to add another
card slot just for invalid cards, where invalid cards cannot be traded
or otherwise used elsewhere in the game.