Add SIGINT handling to properly deallocate memory
Up until now, no memory was being deallocated when SIGINT was thrown.
Note that function prototypes used by event_assign must be defined with
evutil_socket_t as opposed to the int used in my own event callbacks.
This doesn't make a difference because evutil_socket_t is actually
defined as a macro for int in event2/util.h.