Change sizeof calls to use type, not instance
It was bugging me that this wasn't consistent with the rest of the code,
which uses sizeof(struct name) instead of sizeof(instance). Plus, I
personally think it makes more sense unless you have allocated something
on the stack whose size has already been defined, as is the case of the
data variable in buffered_on_read().