Patch a few memory leaks
Memory leak prevention should occur at the lowest possible level. If a
method throws an error, it is entirely responsible to clean up any
associated memory it no longer needs. This is evaluated considering that
on an error, the program immediately stops processing the current
command and moves on to the next one in the packet.