Add start of client GUI
There's quite a bit of code overlap going on already. This will be
rectified in due time when I eventually take out code in the Python bot
that can be used in an external Billionaire library. For now, it will
remain as-is, as only by writing this second client do I see what parts
of the original bot can actually be re-used in other clients.

Also, as I am using Gtk 3.0 Python bindings, I have to use the infamous
  gi.require_version('Gtk', '3.0')
before eventually including Gtk. This means I've had to rewire the pre-
commit hook to ignore E402 which is raised when import statements appear
after code. I'll be keeping all import statements at the top of files,
and I consider the gi.require_version() to be a part of these, despite
being a non-import line of code as seen by Python.