Add margins and disabled buttons
Widgets are disabled using the set_sensitive() method. This allows the
offer button to be grayed out when there's nothing in the player's hand,
as well as allowing the wildcard options to also be grayed out to
represent when the player has these cards in their hand.

Also added some margins for some nice eye room. :)