Split wildcard options up, add abstract CardData
Wildcard options have been split into their own object, the Hand objects
have been renamed to reflect their purpose to only work with commodity
cards. Both of these classes use the new CardData virtual class, which
behaves similarly to a CardLocation in that cards can be added and taken
from it, but also allows the further overriding of implementation-
specific methods. For instance, overriding these methods allows
CommodityData to update the values in the Gtk.ListStore as well.
Similarly, Wildcards updates the sensitivity of its check buttons when
it has the appropriate cards in the hand.