Turn HandDisplay into a frame containing the grid
This allows the entire section to be labelled, which I like for the
clarity. Also updated the name to more accurately reflect the purpose of
the object.

Note in the future I'll have to replace all calls to set_name() with
set_css_name(), as the set_name() is really only for instances of
objects, whilst set_css_name() is for the class as a whole. This is
because set_name() goes into a # identifier, whilst set_css_name()
creates a base element.