Redefine len() for CardData
len() now returns the total amount of cards stored in the CardData
object. This makes a lot more sense, and does not require rewriting its
usage for Wildcards, as this is a CardData object that is only expected
to contain one of each wildcard. It is unlikely that this will be
changed in the future.