Add JSONInterface and CardLocation objects
JSONInterface is an interface that should be used when an object is able
to be written to valid JSON, such as a CommandList or CardLocation.

CardLocation is functionally identical to the client card_location,
except it makes use of the CardData and JSONInterface classes for
modularity and functionality, of course.