Add MessagePasser boilerplate
This object will act as the interface between the GUI and the server. It
takes signals emitted by the GUI and turns them into CommandList
messages that are passed to the server. Likewise, it also reads any
messages sent by the server and passes them through to the GUI.

As of now, it simply responds to the new_offer signal by printing out
the offer object.