Remove super() calls for some Gtk-based objects
This was done because in the future, I will like some of these objects
to inherit from another base class. As a consequence, I cannot user
super() because I haven't figured out how to get it to play nice when
its parent classes need different parameters and have different
footprints.