Move game-handling code into separate file
The server now only deals with setting up ports and the like. To split
up the code, the static structures `billionaire_game` and
`hashed_clients` have had to be moved to `game_state.h` and
`client_hash_table.h` and defined with the `extern` storage class. This
allows both the server to initialise these structures when setting up
the server in `server.c`, and enables the game to interact with these
objects in `billionaire.c`.

Alongside this, includes have been cleaned up so source files now
include their matching header as the first include, before any headers
defined on the path and then any local headers, just to tidy things up a
bit.